fix: update relative import paths after componentname/index.tsx restructure
