+++ apps/tauri/src/renderer/features/support/components/LogsDiagnostics/index.tsx
@@ -2,8 +2,8 @@ import { Button } from '@ajh/ui';
 
 import { useTranslation } from '@/lib/i18n';
 
-import { LogCategoryCard } from './LogCategoryCard';
-import { LogEntry } from './LogEntry';
+import { LogCategoryCard } from '../LogCategoryCard';
+import { LogEntry } from '../LogEntry';
 
 export function LogsDiagnostics() {
   const { t } = useTranslation();
