+++ apps/tauri/src/renderer/features/support/components/PerformanceMemory/index.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from '@/lib/i18n';
 
-import { MemoryBar } from './MemoryBar';
-import { MetricCard } from './MetricCard';
-import { OptimizationCard } from './OptimizationCard';
+import { MemoryBar } from '../MemoryBar';
+import { MetricCard } from '../MetricCard';
+import { OptimizationCard } from '../OptimizationCard';
 
 export function PerformanceMemory() {
   const { t } = useTranslation();
