+++ apps/tauri/src/renderer/routes/resumes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-import { ResumesPage } from './resumes/components/ResumesPage';
+import { ResumesPage } from './components/ResumesPage';
 
 export const Route = createFileRoute('/resumes')({ component: ResumesPage });
